% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STmetabolism.R
\docType{data}
\name{sc}
\alias{sc}
\title{Single-cell example data (sc)}
\format{
A Seurat object with 25314 features and 8944 samples.
}
\usage{
sc
}
\description{
A small example dataset of single-cell RNA-seq data.
}
\details{
Title ST.metabolism Function
}
\keyword{datasets}
