% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STmetabolism.R
\name{ST.metabolism}
\alias{ST.metabolism}
\title{ST.metabolism Function}
\usage{
ST.metabolism(
  st_data,
  sc_data,
  sc_assay = "RNA",
  cell_names = "cell_type",
  int_assay = "traint",
  reduction = "pca",
  intp = TRUE,
  intp_pnt = 5000,
  intp_lin = FALSE,
  nPCs = 30,
  ntree = 1000,
  dist_thresh = 0.55,
  top_spot = 5,
  spot_n = 5,
  repel_r = 20,
  repel_iter = 20,
  keep_model = TRUE,
  method = "VISION",
  imputation = FALSE,
  ncores = 1,
  metabolism.type = "KEGG"
)
}
\arguments{
\item{st_data}{一个V4空间转录组Seurat对象}

\item{sc_data}{一个V4单细胞转录组Seurat对象}

\item{sc_assay}{单细胞的assay对象名称}

\item{cell_names}{metadata中细胞名称的列名}

\item{int_assay}{嵌入空间转录组数据的单细胞数据的assay对象}

\item{reduction}{选择降维的方式}

\item{intp}{如果为TRUE则执行插值}

\item{intp_pnt}{插值点数}

\item{intp_lin}{如果为TRUE则执行线性插值}

\item{nPCs}{PC数量}

\item{ntree}{树数}

\item{dist_thresh}{阈值距离}

\item{top_spot}{一个单元格可以绘制的最大点数}

\item{spot_n}{一个点可以包含的最大单元格数}

\item{repel_r}{排斥半径}

\item{repel_iter}{排斥迭代}

\item{keep_model}{如果为TRUE则返回经过训练的随机森林模型}

\item{method}{进行单细胞代谢评分的方法}

\item{imputation}{是否在代谢评分之前插补其数据}

\item{ncores}{计算单细胞代谢使用的内核数}

\item{metabolism.type}{选择代谢分析的参考数据库}
}
\value{
一个列表，包含STmetbolism对象，celltrek对象和训练后的嵌入数据集
}
\description{
This function integrates single-cell and spatial data for
    metabolic analysis.
}
\examples{
# library(Seurat,lib.loc = 'D:\Program Files\R\R-4.4.1\seurat_lib')
    #V4版本的Seurat
library(CellTrek)
library(scMetabolism)
library(dplyr)
# library(ggplot2)
# library(rsvd)
library(STmetabolism)
# 需要输入V4 Seurat创建的单细胞对象和空转对象
head(sc)
sc
head(st)
st
ST_met <- ST.metabolism(st_data = st, sc_data = sc)
# 查看训练后的嵌入数据
DimPlot(ST_met$traint, group.by = "type")
# celltrek对象可以直接按照celltrek的教程运行后续步骤
CellTrek::celltrek_vis(
  ST_met$celltrek@meta.data \%>\%
    dplyr::select(coord_x, coord_y, cell_type:id_new),
  ST_met$celltrek@images$anterior1@image,
  ST_met$celltrek@images$anterior1@scale.factors$lowres
)
# 以celltrek的点为基础绘制切片的代谢活性分布图
SpatialFeaturePlot(ST_met$STmetbolism_obj,
  features = c("Glycolysis / Gluconeogenesis"), pt.size.factor = 2.5
)

}
