% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STmetabolism.R
\docType{data}
\name{st}
\alias{st}
\title{Spatial example data (st)}
\format{
A Seurat object with 36601 features and 2261 samples.
}
\usage{
st
}
\description{
A small example dataset of spatial transcriptomics data.
}
\keyword{datasets}
